refactor: move node/startup to services
