+++ cmd/cartesi-rollups-validator/root/root.go
@@ -12,8 +12,8 @@ import (
 	"time"
 
 	"github.com/cartesi/rollups-node/internal/config"
-	"github.com/cartesi/rollups-node/internal/node/startup"
 	"github.com/cartesi/rollups-node/internal/repository"
+	"github.com/cartesi/rollups-node/internal/services/startup"
 	"github.com/cartesi/rollups-node/internal/validator"
 	"github.com/spf13/cobra"
 )
